Consolidate fixed math functions into rt_fixed.c
