+++ rott/_rt_play.h
@@ -20,7 +20,6 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #ifndef _rt_playr_private
 #define _rt_playr_private
 
-#include "watcom.h"
 //double dTopYZANGLELIMIT; in winrott.c
 #define TopYZANGLELIMIT  (44*FINEANGLES/360)//bna added
 #define YZANGLELIMIT  (80*FINEANGLES/360)//bna--(30*FINEANGLES/360)
