+++ rott/engine.c
@@ -18,7 +18,6 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 */
 #include "rt_def.h"
-#include "watcom.h"
 #include "engine.h"
 #include "_engine.h"
 #include "rt_draw.h"
