+++ rott/rt_actor.c
@@ -26,13 +26,11 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "rt_door.h"
 #include "rt_ted.h"
 #include "rt_draw.h"
-#include "watcom.h"
 #include "z_zone.h"
 #include "w_wad.h"
 #include "lumpy.h"
 #include "gmove.h"
 #include "states.h"
-#include "rt_sqrt.h"
 #include "rt_stat.h"
 #include "sprites.h"
 #include "rt_actor.h"
@@ -2036,7 +2034,7 @@ void T_Spring(objtype*ob)
       if (!temp->momentumz)
          {
          op = FixedMul(GRAVITY,(temp->z-5)<<16) << 1;
-         temp->momentumz = -FixedSqrtHP(op);
+         temp->momentumz = -FixedSqrt(op);
          SD_PlaySoundRTP(SD_SPRINGBOARDSND,ob->x,ob->y);
          }
       }
@@ -4616,7 +4614,7 @@ void HeatSeek(objtype*ob)
        dx = tactor->x - ob->x;
        dy = ob->y - tactor->y;
        dz = ob->z - tactor->z;
-       //xydist = FixedSqrtHP((FixedMul(dx,dx) + FixedMul(dy,dy))>>16);
+       //xydist = FixedSqrt((FixedMul(dx,dx) + FixedMul(dy,dy))>>16);
        xydist = FindDistance(dx,dy);
        angle = atan2_appx(dx,dy);
        adjust = (ob->obclass == p_godballobj)?(GODHAPT):(HAAPT);
@@ -6288,7 +6286,7 @@ movement_status CheckOtherActors(objtype*ob,int tryx,int tryy,int tryz)
             {
                {
                op = (FixedMul((int)GRAVITY,(int)((ob->z-10)<<16))<<1);
-               ob->momentumz = -FixedSqrtHP(op);
+               ob->momentumz = -FixedSqrt(op);
                }
                SD_PlaySoundRTP(SD_SPRINGBOARDSND,listrover->x,listrover->y);
                NewState(listrover,&s_spring2);
@@ -9111,7 +9109,7 @@ void T_NME_Explode(objtype*ob)
 	  head = (objtype*)(ob->whatever);
 
 	  op = FixedMul(GRAVITY,(head->z-25)<<16) << 1;
-	  head->momentumz = -FixedSqrtHP(op);
+	  head->momentumz = -FixedSqrt(op);
 	  head->momentumx = (GameRandomNumber("NME head momx",0) << 2);
 	  head->momentumy = (GameRandomNumber("NME head momy",0) << 2);
      head->hitpoints = 0;
