+++ rott/rt_build.c
@@ -21,7 +21,6 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 #include "rt_def.h"
 #include <string.h>
-#include "watcom.h"
 #include <stdio.h>
 #include <stdlib.h>
