+++ rott/rt_draw.c
@@ -23,7 +23,6 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "rt_def.h"
 #include <string.h>
 
-#include "watcom.h"
 #include "sprites.h"
 #include "rt_actor.h"
 #include "rt_stat.h"
