+++ rott/rt_map.c
@@ -48,7 +48,6 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "rt_menu.h"
 #include "rt_net.h"
 #include "rt_str.h"
-#include "watcom.h"
 
 //===========================================================================
