+++ rott/rt_scale.c
@@ -18,7 +18,6 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 */
 #include "rt_def.h"
-#include "watcom.h"
 #include <stdio.h>
 #include <string.h>
