+++ rott/rt_sound.c
@@ -30,7 +30,6 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "rt_playr.h"
 #include "rt_util.h"
 #include "rt_rand.h"
-#include "watcom.h"
 #include <stdio.h>
 #include <stdlib.h>
