+++ rott/rt_ted.c
@@ -29,7 +29,6 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include <errno.h>
 
 #include "states.h"
-#include "watcom.h"
 #include "rt_ted.h"
 #include "_rt_ted.h"
 #include "w_wad.h"
