+++ rott/rt_util.c
@@ -31,7 +31,6 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include <stdlib.h>
 #include <sys/stat.h>
 #include <time.h>
-#include "watcom.h"
 #include "_rt_util.h"
 #include "rt_util.h"
 #include "isr.h"
