+++ rott/rt_view.c
@@ -39,7 +39,6 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "rt_battl.h"
 #include "rt_floor.h"
 #include "rt_str.h"
-#include "watcom.h"
 #include "develop.h"
 
 #define LIGHTNINGLEVEL 4
